refactor: consolidate divergent types across apps into shared libraries
- Move notification enums (NotificationType, NotificationPriority,
  NotificationChannel, RelatedEntityType) and helper functions into
  lib.types as the single source of truth; all 3 apps now re-export
  from there instead of maintaining identical copies.

- Refactor app.client types to import User, Pet, Application,
  discovery types, and PaginatedResponse from their canonical library
  packages (lib.auth, lib.pets, lib.applications, lib.discovery,
  lib.types) instead of duplicating 400+ lines of hand-written
  interfaces.

- Refactor app.admin types to import auth types from lib.auth and
  UserRole from lib.types, eliminating narrower local type unions
  that would silently reject valid role values.

- Unify PaginatedResponse across lib.pets and lib.rescue to re-export
  from lib.types, eliminating 3 separate identical definitions.

- Normalise admin pagination from legacy `pages` field to canonical
  `totalPages` (matching the shared type contract).

All 1,213 tests pass; type-check clean across all 3 apps.
