+++ app.client/package.json
@@ -24,6 +24,7 @@
     "@adopt-dont-shop/lib.api": "*",
     "@adopt-dont-shop/lib.applications": "*",
     "@adopt-dont-shop/lib.auth": "*",
+    "@adopt-dont-shop/lib.types": "*",
     "@adopt-dont-shop/lib.chat": "*",
     "@adopt-dont-shop/lib.components": "*",
     "@adopt-dont-shop/lib.dev-tools": "*",
