+++ lib.pets/package.json
@@ -37,6 +37,7 @@
   "license": "MIT",
   "dependencies": {
     "@adopt-dont-shop/lib.api": "*",
+    "@adopt-dont-shop/lib.types": "*",
     "@types/node": "^20.19.0",
     "zod": "^4.4.3"
   },
