+++ lib.pets/src/types/index.ts
@@ -10,18 +10,8 @@ export type {
   PetUpdateData,
 } from '../schemas';
 
-// PaginatedResponse is generic and cannot be derived from a Zod schema directly
-export type PaginatedResponse<T> = {
-  data: T[];
-  pagination: {
-    page: number;
-    limit: number;
-    total: number;
-    totalPages: number;
-    hasNext: boolean;
-    hasPrev: boolean;
-  };
-};
+// ADS-262: response envelopes are owned by @adopt-dont-shop/lib.types.
+export type { PaginatedResponse } from '@adopt-dont-shop/lib.types';
 
 export type PetsServiceConfig = {
   apiUrl?: string;
